Code clean up
1. Removed all references to jMonkey
2. Removed ui/helper.py as all the information was outdated and replaced by the readme
3. Removed code from ui/__init__.py referencing a material viewer (& editor?)
 - I got scared because it looked like quite a lot of code could get removed.
  - ui/helper.py & ui/material.py were removed. The code in material.py was moved to unused.py
  - Potentially the following files could be removed. I don't know enough about how the plugin is used by other people to be sure. It removes quite a lot of code, which to me looks unused due to API changes and unmaintained features:
    - properties.py
    - game_logic.py
    - terrain.py
